Replace Thread.isAlive with Thread.is_alive
Thread.isAlive method has been deprecated in Python 3.8 and got removed
in 3.9 [1].

On some distributions like Arch Linux, Kodi is run with newest Python
(as for this day 3.9.3). As a result this addon is broken because of:

    ERROR <general>: AttributeError
    ERROR <general>: :
    ERROR <general>: 'FileGetter' object has no attribute 'isAlive'
    ERROR <general>:

This commit fixes the issue.
is_alive naming spelling has been already introduced in 2.6 [2] so
hopefully this change is safe for Kodi Leia.

[1]: https://bugs.python.org/issue37804
[2]: https://devdocs.io/python~2.7/library/threading#threading.Thread.is_alive